+++ client/cypress/e2e/example.cy.ts
@@ -1,5 +1,3 @@
-// https://on.cypress.io/api
-
 describe('My First Test', () => {
   it('visits the app root url', () => {
     cy.visit('/')
