+++ client/src/components/BJButton.cy.ts
@@ -1,0 +1,8 @@
+import BJButton from './BJButton.vue'
+
+describe('<BJButton />', () => {
+  it('mounts', () => {
+    // see: https://on.cypress.io/mounting-vue
+    cy.mount(BJButton)
+  })
+})
