+++ client/tsconfig.app.json
@@ -1,7 +1,7 @@
 {
   "extends": "@vue/tsconfig/tsconfig.dom.json",
   "include": ["env.d.ts", "src/**/*", "src/**/*.vue"],
-  "exclude": ["src/**/__tests__/*"],
+  "exclude": ["src/**/*.spec.ts","src/**/*.cy.ts"],
   "compilerOptions": {
     "composite": true,
     "noEmit": true,
