+++ client/tsconfig.json
@@ -9,7 +9,7 @@
     },
     {
       "path": "./tsconfig.vitest.json"
-    }
+    },
   ],
   "compilerOptions": {
     "module": "NodeNext"
